% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/QFSTS.R
\name{QFSTS.func}
\alias{QFSTS.func}
\title{Main function for the quantile feature selection time series.}
\usage{
QFSTS.func(
  Y,
  X.star = NULL,
  STmodel = NULL,
  ki = NULL,
  pii = NULL,
  b = NULL,
  kapp = 0.1,
  R2 = 0.8,
  v0 = NULL,
  v = 0.01,
  ss = 0.01,
  tau = NULL,
  Phi = NULL,
  mc = 500,
  burn = 50
)
}
\arguments{
\item{Y}{A (n ∗ m)-dimensional matrix containing multiple target series, where n is thenumber of observations and m
is the number of target series.}

\item{X.star}{A (n ∗ K)-dimensional matrix containing all candidate predictor series for each target series.
K = Simga k_i is the number of all candidate predictors for all target}

\item{STmodel}{A state space model of STmodel class returned by tsc.setting.}

\item{ki}{A vector of integer values denoting the acumulated number of predictors for}

\item{pii}{A vector describing the prior inclusion probability of each candidate predictor.}

\item{b}{NULL or a vector describing the prior means of regression coefficients. The default value is NULL.}

\item{kapp}{A scalar value describing the number of observations worth of weight on the prior mean vector. The default value is 0.01.}

\item{R2}{A numerical value taking value in [0, 1], describing the expected percentage of variation of Y to be explained by the model. The default value is 0.8.}

\item{v0}{A numerical value describing the prior degree of freedom of the inverse Wishart distribution for sigma_epsilon.}

\item{v}{A numerical value describing the prior degree of freedom of the inverse Wishart distribution for (Σµ, Σδ, Στ , Σω). The default value is 0.01}

\item{ss}{A numerical value describing the prior scale matrix of the inverse Wishart distribution for (Σµ, Σδ, Στ , Σω). The default value is 0.01.}

\item{tau}{A vecotr containing the quantile values.}

\item{Phi}{A m × m-dimensional diagonal matrix of means of error terms.}

\item{mc}{A positive integer giving the desired number of MCMC draws. The default value is 500}

\item{burn}{A positive integer giving the number of initial MCMC draws to be discarded. The default value is 50.}
}
\value{
An object of QFSTS class
}
\description{
Main function for the quantile feature selection time series.
}
\examples{
Ytrain = cbind(rnorm(100), rnorm(100))
QFSTS.func(Ytrain)
}
