% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/3_series_eg1.R
\name{sim.data.func}
\alias{sim.data.func}
\title{Title This function simulates 3 time series data to be used}
\usage{
sim.data.func(n, tau, corr)
}
\arguments{
\item{n}{number of observations}

\item{tau}{A vecotr containing the quantile values.}

\item{corr}{correlation value}
}
\value{
a dataset to train
}
\description{
Title This function simulates 3 time series data to be used
}
\examples{
tau1<-c(0.1,0.1,0.1)
count=100
for (n in count){
nam <- paste("simdata_3series_tau1_",n, sep = "")
assign(nam, sim.data.func(n,tau1,corr=0.4))}


}
