% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tsc.setting.QFSTS.R
\name{tsc.setting}
\alias{tsc.setting}
\title{Specification of time series components}
\usage{
tsc.setting(
  Ytrain,
  mu = NULL,
  rho = NULL,
  S = NULL,
  vrho = NULL,
  lambda = NULL
)
}
\arguments{
\item{Ytrain}{The multivariate time series to be modeled.}

\item{mu}{A vector of logic values indicating whether to include a local trend for each target series.}

\item{rho}{A vector of numerical values taking values in [0, 1], describing the learning rates at which the local trend is updated for each target series. The value 0 in the j-th entry indicates that the j-th target series does not include slope of trend.}

\item{S}{A vector of integer values representing the number of seasons to be modeled for each target series. The value 0 in the j-th entry indicates that the j-th target series does not include the seasonal component.}

\item{vrho}{A vector of numerical values taking values in [0, 1], describing a damping factor for each target series. The value 0 in the j-th entry indicates that the j-th target series does not include the cycle component.}

\item{lambda}{A vector of numerical values, whose entries equal to 2π/q with q being a period such that 0 < λ < π, describing the frequency.}
}
\value{
Structural time series model
}
\description{
Specification of time series components
}
\examples{
tau1<-c(0.1,0.1,0.1)
count=100
for (n in count){
nam <- paste("simdata_3series_tau1_",n, sep = "")
assign(nam, sim.data.func(n,tau1,corr=0.4))}
Ytrain<-as.matrix(simdata_3series_tau1_100[,1:3])

STmodel<-tsc.setting(Ytrain,mu=c(1,1,1), rho=c(0.6,0.3,0.1), S=c(100,70,40))


}
